+++ src/express.ts
@@ -5,43 +5,67 @@ import { AckeeLogger } from '.';
 
 const errorSymbol = Symbol.for('error');
 
+type AckeeRequest = Request & { _startAt?: [number, number]; ackId?: string };
+type AckeeResponse = Response & { _startAt?: [number, number]; time?: string; out?: object; [errorSymbol]?: any };
+
 export type AckeeLoggerExpressMiddleware = (
     this: AckeeLogger,
-    req: Request & { _startAt?: [number, number]; ackId?: string },
-    response: Response & { _startAt?: [number, number]; time?: string; out?: object; [errorSymbol]?: any },
+    req: AckeeRequest,
+    response: AckeeResponse,
     next: any
 ) => void;
 
+const expressOnHeaders = (req: AckeeRequest, res: AckeeResponse) => () => {
+    res._startAt = process.hrtime();
+    const diffFromSeconds = (res._startAt[0] - req._startAt![0]) * 1e3;
+    const diffFromNanoseconds = (res._startAt[1] - req._startAt![1]) * 1e-6;
+    const ms = diffFromSeconds + diffFromNanoseconds;
+    res.time = ms.toFixed(3);
+};
+
+const expressOnFinished = (logger: AckeeLogger, req: AckeeRequest) => (_err: Error | null, res: AckeeResponse) => {
+    const error = res[errorSymbol];
+    const userAgent = req.headers['user-agent'];
+    const reqOut = `${res.statusCode} ${req.method} ${req.originalUrl} ${res.time} ms ${userAgent ? userAgent : ''}`;
+    const standardOutput = {
+        data: {
+            req,
+            res,
+            ackId: req.ackId,
+        },
+        message: `${reqOut} - Standard output`,
+    };
+    const errorOutput = {
+        data: { ...standardOutput.data, error },
+        message: `${reqOut} - Error handler at the end of app`,
+    };
+    const serverError = res.statusCode >= 500;
+
+    const logFunction = error || serverError ? logger.error : res.out ? logger.info : logger.trace;
+    const output = error ? errorOutput : standardOutput;
+
+    logFunction.call(logger, output.data, output.message);
+};
+
 const expressMiddleware: RequestHandler = function(
     this: AckeeLogger,
-    req: Request & { _startAt?: [number, number]; ackId?: string },
-    response: Response & { _startAt?: [number, number]; time?: string; out?: object; [errorSymbol]?: any },
+    req: AckeeRequest,
+    response: AckeeResponse,
     next: any
 ) {
-    const reqIn = `--- ${req.method} ${req.originalUrl} ${req.headers['user-agent']}`;
-    this.debug({ req, ackId: req.ackId }, `${reqIn} - Request accepted`);
+    const userAgent = req.headers['user-agent'];
+    const reqIn = `--- ${req.method} ${req.originalUrl} ${userAgent ? userAgent : ''}`;
+    if (this.options.ignoredHttpMethods && this.options.ignoredHttpMethods.includes(req.method)) {
+        // entire method skipped - left here for BC
+        return next();
+    }
+    if (!this.options.skip || !this.options.skip(req)) {
+        // if request not skipped
+        this.debug({ req, ackId: req.ackId }, `${reqIn} - Request accepted`);
+    }
     req._startAt = process.hrtime();
-    onHeaders(response, () => {
-        response._startAt = process.hrtime();
-        const diffFromSeconds = (response._startAt[0] - req._startAt![0]) * 1e3;
-        const diffFromNanoseconds = (response._startAt[1] - req._startAt![1]) * 1e-6;
-        const ms = diffFromSeconds + diffFromNanoseconds;
-        response.time = ms.toFixed(3);
-    });
-    onFinished(response, (_err, res) => {
-        const error = res[errorSymbol];
-        const reqOut = `${res.statusCode} ${req.method} ${req.originalUrl} ${res.time} ms ${req.headers['user-agent']}`;
-        if (this.options.ignoredHttpMethods && this.options.ignoredHttpMethods.includes(req.method)) {
-            return;
-        }
-        if (error) {
-            this.error({ error, req, res, ackId: req.ackId }, `${reqOut} - Error handler at the end of app`);
-        } else if (res.out) {
-            this.debug({ req, res, ackId: req.ackId }, `${reqOut} - Standard output`);
-        } else {
-            this.info({ req, res, ackId: req.ackId }, `${reqOut} - Standard output`);
-        }
-    });
+    onHeaders(response, expressOnHeaders(req, response));
+    onFinished(response, expressOnFinished(this, req));
     next();
 };